+++ Dialogflow-18/index.js
@@ -1,76 +1,11 @@
-const dialogflow = require('@google-cloud/dialogflow');
-require('dotenv').config();
+const express = require("express");
 
-const projectId = process.env.projectId;
-console.log(projectId)
-const sessionId = '123456';
-const queries = ['whats your age']
-
-const languageCode = 'en';
-
-const sessionClient = new dialogflow.SessionsClient({
-    private_key: process.env.private_key,
-    client_email: process.env.client_email,
-});
-
-async function detectIntent(
-    projectId,
-    sessionId,
-    query,
-    contexts,
-    languageCode
-) {
-    // The path to identify the agent that owns the created intent.
-    const sessionPath = sessionClient.projectAgentSessionPath(
-        projectId,
-        sessionId
-    );
-
-    // The text query request.
-    const request = {
-        session: sessionPath,
-        queryInput: {
-            text: {
-                text: query,
-                languageCode: languageCode,
-            },
-        },
-    };
-
-    if (contexts && contexts.length > 0) {
-        request.queryParams = {
-            contexts: contexts,
-        };
-    }
-
-    const responses = await sessionClient.detectIntent(request);
-    return responses[0];
-}
-
-async function executeQueries(projectId, sessionId, queries, languageCode) {
-    // Keeping the context across queries let's us simulate an ongoing conversation with the bot
-    let context;
-    let intentResponse;
-    for (const query of queries) {
-        try {
-            console.log(`Sending Query: ${query}`);
-            intentResponse = await detectIntent(
-                projectId,
-                sessionId,
-                query,
-                context,
-                languageCode
-            );
-            console.log('Detected intent');
-            console.log(intentResponse.queryResult.fulfillmentText);
-            // Use the context from this response for next queries
-            context = intentResponse.queryResult.outputContexts;
-        } catch (error) {
-            console.log(error);
-        }
-    }
-}
+const app = express();
+app.get("/", function(req, res) {
+    res.send("hello")
+})
 
 
-
-executeQueries(projectId, sessionId, queries, languageCode);
+app.listen(5000, function() {
+    console.log("server running")
+})
